+++ NIRS/Additional Calculated Values/PseudoViscosityMechanism.cs
@@ -15,7 +15,8 @@ namespace NIRS.Additional_calculated_values
             double NablaV = wc.Nabla(PN.v, n, k);
             if (NablaV < 0)
             {
-                return Math.Pow(constP.mu0, 2) * Math.Pow(constP.h, 2) * g[n + 0.5][k].ro * Math.Pow(NablaV, 2);
+                return (double)
+                    (Math.Pow(constP.mu0, 2) * Math.Pow(constP.h, 2) * g[n + 0.5][k].ro * Math.Pow(NablaV, 2));
             }
             else
                 return 0;
